+++ ExpenseTracker/src/service/transactionService.js
@@ -5,10 +5,12 @@ const BASE_URL = `${import.meta.env.VITE_BACKEND_URL}/transaction`;
 const api = axios.create({
     baseURL: BASE_URL 
 });
+
+const token = localStorage.getItem("token")
   
 const getAllTransaction = async (page, limit) => {
     try {
-        const res = await api.get(`/${page}/${limit}`); 
+        const res = await api.get(`/${page}/${limit}`, {headers:{Authorization:`Bearer ${token}`}}); 
         console.log('res', res.data);
         return res.data;
     } catch (error) {
@@ -18,7 +20,7 @@ const getAllTransaction = async (page, limit) => {
 
 const getGraphDetails = async () => {
     try {
-        const res = await api.get(`/graph-details`); 
+        const res = await api.get(`/graph-details`, {headers:{Authorization:`Bearer ${token}`}}); 
         console.log('res', res.data);
         return res.data;
     } catch (error) {
@@ -28,7 +30,7 @@ const getGraphDetails = async () => {
 
 const getOneTransaction = async (id) => {
     try {
-        const res = await api.get(`/${id}`); 
+        const res = await api.get(`/${id}`, {headers:{Authorization:`Bearer ${token}`}}); 
         return res.data;
     } catch (error) {
         console.log('Error fetching data:', error);
@@ -46,7 +48,7 @@ const updateTransaction = async (id, data) => {
 
 const deleteTrasnaction = async (id) => {
     try {
-        const res = await api.delete(`/${id}`);
+        const res = await api.delete(`/${id}`, {headers:{Authorization:`Bearer ${token}`}});
         return res.data;
     } catch (error) {
         console.log(error)
@@ -55,7 +57,7 @@ const deleteTrasnaction = async (id) => {
 
 const createTrasnaction = async (data) => {
     try {
-        const res = await api.post(`/`, data);
+        const res = await api.post(`/`, data, {headers:{Authorization:`Bearer ${token}`}});
         return res.data;
     } catch (error) {
         console.log(error)
